+++ sablecc/grammar/postfix/semantics/visitors/SemanticVisitor.java
@@ -136,6 +136,28 @@ public class SemanticVisitor extends DepthFirstAdapter {
     }
 
     @Override
+    public void caseAFunctionDeclarationDcl(AFunctionDeclarationDcl node) {
+        inAFunctionDeclarationDcl(node);
+        if (node.getType() != null) {
+            node.getType().apply(this);
+        }
+        if (node.getKwFunction() != null) {
+            node.getKwFunction().apply(this);
+        }
+        if (node.getId() != null) {
+            node.getId().apply(this);
+        }
+        //! uh oh
+        symbolTable = symbolTable.getFunctionSymbolTable(node.getId().getText());
+        if (node.getFunctionParam() != null) {
+            node.getFunctionParam().apply(this);
+        }
+        if (node.getStmts() != null) {
+            node.getStmts().apply(this);
+        }
+        outAFunctionDeclarationDcl(node);
+    }
+    @Override
     public void inAFunctionCallFunctionCall(AFunctionCallFunctionCall node) {
         // funktionsparametre
