+++ sablecc/grammar/semanticTest.txt
@@ -42,7 +42,7 @@ while i < 3 {
 # bool g = 5 + 6;
 # bool x = 2 < 3;
 
-string function foo(int a, int b) {return a+b ; };
+string function foo(int a, int d) {return a+d ; };
 bool function bar(int c, bool b) { if b {return foo(c,2);} else {return -1;}; };
 
 # bar(a,true);
